+++ src/components/ui/Navbar.tsx
@@ -9,6 +9,7 @@ export const Navbar: FC = () => {
       <ActiveLink text={'About'} href={'/about'} />
       <ActiveLink text={'Users'} href={'/users'} />
       <ActiveLink text={'Notifications'} href={'/users/notifications'} />
+      <ActiveLink text={'Transactions'} href={'/transactions'} />
       <ActiveLink text={'Friends'} href={'/friends'} />
     </nav>
   );
